Decompose CLI into smaller modules
Topic: decompose-cli
